+++ module/item/item-requires-roll.mjs
@@ -2,6 +2,7 @@ import { HEROSYS } from "../herosystem6e.mjs";
 import { HeroRoller } from "../utility/dice.mjs";
 import { calculateDicePartsForItem } from "../utility/damage.mjs";
 import { overrideCanAct } from "../settings/settings-helpers.mjs";
+import { tokenEducatedGuess, whisperUserTargetsForActor } from "../utility/util.mjs";
 
 const backgroundSkillKeys = Object.freeze({
     // matches RaR OPTION to XMLID
@@ -275,6 +276,66 @@ export async function isActivatedForThisUse(item, options = {}) {
         return true;
     }
 
+    const actor = item.actor;
+    const token = options.token ?? tokenEducatedGuess({ actor });
+    const speaker = ChatMessage.getSpeaker({ actor, token });
+
+    // Sectional Defense is new, so putting a safety TRY/CATCH around the new code
+    try {
+        function determineSectionalDefenses(comment, hitLocationNum) {
+            // We should always have a hitLocationNum, being paranoid
+            if (!hitLocationNum) {
+                console.error(`hitLocationNum was not found`);
+                return null;
+            }
+
+            // Expecting comment to contain hit location details.  For example "locations 1-18".
+            // TODO: Support "locations 3-5, 9-14, 16-18"
+            // TODO: Move this to be an item getter so we can use it in HeroValidation to determine if the RAR 8- 14-, etc are correct.
+            const hitLocationMatch = comment?.match(/locations (\d+)-(\d+)/i);
+            if (hitLocationMatch) {
+                return {
+                    // lower: hitLocationMatch[1], // lower/upper are not used and should be array of ranges if we want to keep
+                    // upper: hitLocationMatch[2],
+                    success: hitLocationNum >= hitLocationMatch[1] && hitLocationNum <= hitLocationMatch[2],
+                };
+            }
+            console.warn(`Sectional Defense location matching failed`);
+            return null;
+        }
+
+        // Sectional Defense will bypass the roll
+        if (options.damageRoller) {
+            const activationRoll =
+                item.modifiers.find((o) => o.XMLID === "ACTIVATIONROLL") ?? item.findModsByXmlid("EVERYPHASE")?.parent;
+            if (activationRoll) {
+                const sectionalDefenses = determineSectionalDefenses(
+                    activationRoll.COMMENTS,
+                    options.damageRoller.getHitLocation().num,
+                );
+
+                if (sectionalDefenses) {
+                    // ChatMessage
+                    const chatData = {
+                        style: CONST.CHAT_MESSAGE_STYLES.IC,
+                        author: game.user._id,
+                        content: `The sectional defense from <b>${item.name}</b> ${sectionalDefenses.success ? "successfully applied" : "failed to apply"}`,
+                        speaker: speaker,
+                        whisper: whisperUserTargetsForActor(actor),
+                    };
+                    await ChatMessage.create(chatData);
+                    return sectionalDefenses.success;
+                } else {
+                    console.warn(
+                        `Check for Sectional Defense failed, expected "locations x-y" in the COMMENTS. Will use standard activation roll instead.`,
+                    );
+                }
+            }
+        }
+    } catch (e) {
+        console.error(`Unexpected sectional defense error`, e);
+    }
+
     // FIXME: This doesn't support 2 RAR. See https://github.com/dmdorman/hero6e-foundryvtt/issues/3873
 
     // todo why OPTION and not OPTIONID to look for background skills (OPTIONID can't be altered by user)
@@ -309,10 +370,6 @@ export async function isActivatedForThisUse(item, options = {}) {
                     <br>Put the name of the ${typeOfRollRequired} into the Options or Comments of the Requires A Roll modifier for the best results.`,
                 );
                 if (!overrideCanAct) {
-                    const actor = item.actor;
-                    const token = actor.token;
-                    const speaker = ChatMessage.getSpeaker({ actor: actor, token });
-                    speaker.alias = actor.name;
                     const overrideKeyText = game.keybindings.get(HEROSYS.module, "OverrideCanAct")?.[0].key;
 
                     const chatData = {
@@ -408,10 +465,6 @@ export async function isActivatedForThisUse(item, options = {}) {
         succeeded = true;
         cardHtml += `<p>Succeeded roll because ${game.user.name} used <b>${overrideKeyText}</b> key to override.</p>`;
     }
-
-    const actor = item.actor;
-    const token = actor.token;
-    const speaker = ChatMessage.getSpeaker({ actor: actor, token });
 
     if (!succeeded && options.resourcesUsedDescription) {
         cardHtml += `Spent ${options.resourcesUsedDescription}.`;